requirements.txt
streamlit
openai
pyyaml